+++ src/fixedarray_funcs.c
@@ -23,28 +23,28 @@ void fa_freer(FixedArray *const fa, void (*const f)(data*)) {
 }
 
 void fa_clear(FixedArray *const fa, void (*const f)(data*)) {
-	const unsigned int s = fa_size(fa);
+	const size_t s = fa_size(fa);
 	if(f) {
 		fa_each(fa, f);
 	}
-	for(unsigned int i = 0; i < s; ++i) {
+	for(size_t i = 0; i < s; ++i) {
 		fa_unset(fa, i);
 	}
 }
 
-unsigned int fa_count(const FixedArray *const fa) {
-	const unsigned int s = fa_size(fa);
-	unsigned int n = 0;
-	for(unsigned int i = 0; i < s; ++i) {
+size_t fa_count(const FixedArray *const fa) {
+	const size_t s = fa_size(fa);
+	size_t n = 0;
+	for(size_t i = 0; i < s; ++i) {
 		if(fa_get(fa, i))
 			++n;
 	}
 	return n;
 }
 
-int fa_put(FixedArray *const fa, data *const item) {
-	const unsigned int s = fa_size(fa);
-	for(unsigned int i = 0; i < s; ++i)
+ssize_t fa_put(FixedArray *const fa, data *const item) {
+	const size_t s = fa_size(fa);
+	for(size_t i = 0; i < s; ++i)
 		if(!fa_get(fa, i)) {
 			fa_set(fa, i, item);
 			return i;
@@ -52,7 +52,7 @@ int fa_put(FixedArray *const fa, data *const item) {
 	return -1;
 }
 
-data *fa_swap(FixedArray *const fa, const unsigned int i, data *const e) {
+data *fa_swap(FixedArray *const fa, const size_t i, data *const e) {
 	data *const d = fa_get(fa, i); /* this sets errno */
 	if(errno) {
 		return NULL;
@@ -61,16 +61,16 @@ data *fa_swap(FixedArray *const fa, const unsigned int i, data *const e) {
 		return d;
 	}
 }
-extern data *fa_unset(FixedArray*, unsigned int);
+extern data *fa_unset(FixedArray*, size_t);
 
 
 data *fa_cond(const FixedArray *fa, bool (*f)(const data*, const data*), const data *v) {
-	const unsigned int s = fa_size(fa);
+	const size_t s = fa_size(fa);
 	data *e;
 	if(!f) {
 		f = default_equals;
 	}
-	for(unsigned int i = 0; i < s; ++i) {
+	for(size_t i = 0; i < s; ++i) {
 		e = fa_get(fa, i);
 		if(f(e, v)) {
 			errno = 0;
@@ -82,11 +82,11 @@ data *fa_cond(const FixedArray *fa, bool (*f)(const data*, const data*), const d
 }
 
 data *fa_remove(FixedArray *fa, bool (*f)(const data*, const data*), const data *const v) {
-	const unsigned int s = fa_size(fa);
+	const size_t s = fa_size(fa);
 	if(!f) {
 		f = default_equals;
 	}
-	for(unsigned int i = 0; i < s; ++i) {
+	for(size_t i = 0; i < s; ++i) {
 		if(f(fa_get(fa, i), v)) {
 			errno = 0;
 			return fa_unset(fa, i);
@@ -97,8 +97,8 @@ data *fa_remove(FixedArray *fa, bool (*f)(const data*, const data*), const data 
 }
 
 void fa_each(FixedArray *const fa, void (*const f)(data*)) {
-	const unsigned int s = fa_size(fa);
-	for(unsigned int i = 0; i < s; ++i) {
+	const size_t s = fa_size(fa);
+	for(size_t i = 0; i < s; ++i) {
 		data *const item = fa_get(fa, i);
 		if(item)
 			f(item);
@@ -106,13 +106,13 @@ void fa_each(FixedArray *const fa, void (*const f)(data*)) {
 }
 
 void fa_printf(const FixedArray *const fa, void (*p)(const data*)) {
-	const unsigned int s = fa_size(fa);
+	const size_t s = fa_size(fa);
 	printf("[");
 	if(s) {
 		if(!p)
 			p = default_print_item;
 		p(fa_get(fa, 0));
-		for(unsigned int i = 1; i < s; ++i) {
+		for(size_t i = 1; i < s; ++i) {
 			printf(", ");
 			p(fa_get(fa, i));
 		}
