+++ src/linkedlist_funcs.c
@@ -20,12 +20,12 @@ static void _printitem(const data *const e) {
 
 void ll_each(LinkedList *const ll, void (*const f)(data*)) {
 	data *item;
-	for(unsigned int i = 0; (item = ll_get(ll, i)); ++i) {
+	for(size_t i = 0; (item = ll_get(ll, i)); ++i) {
 		f(item);
 	}
 }
 
-data *ll_swap(LinkedList *const ll, const unsigned int i, data *const d) {
+data *ll_swap(LinkedList *const ll, const size_t i, data *const d) {
 	data *const e = ll_get(ll, i);
 	if(errno) {
 		return NULL;
@@ -35,8 +35,9 @@ data *ll_swap(LinkedList *const ll, const unsigned int i, data *const d) {
 }
 
 
-data *ll_cond(const LinkedList *const ll, const data *const e, bool (*f)(const data*, const data*)) {
-	const unsigned int l = ll_len(ll);
+data *ll_cond(const LinkedList *const ll, const data *const e,
+              bool (*f)(const data*, const data*)) {
+	const size_t l = ll_len(ll);
 	data *item;
 	if(!f) {
 		if(!e) {
@@ -45,7 +46,7 @@ data *ll_cond(const LinkedList *const ll, const data *const e, bool (*f)(const d
 		}
 		f = _equals;
 	}
-	for(unsigned int i = 0; i < l; ++i) {
+	for(size_t i = 0; i < l; ++i) {
 		item = ll_get(ll, i);
 		if(f(item, e))
 			return item;
@@ -55,12 +56,13 @@ data *ll_cond(const LinkedList *const ll, const data *const e, bool (*f)(const d
 }
 
 
-data *ll_remove(LinkedList *const ll, const data *const e, bool (*f)(const data*, const data*)) {
+data *ll_remove(LinkedList *const ll, const data *const e,
+                bool (*f)(const data*, const data*)) {
 	if(!f) {
 		f = _equals;
 	}
-	const unsigned int s = ll_len(ll);
-	for(unsigned int i = 0; i < s; ++i) {
+	const size_t s = ll_len(ll);
+	for(size_t i = 0; i < s; ++i) {
 		if(f(ll_get(ll, i), e)) {
 			errno = 0;
 			return ll_drop(ll, i);
@@ -77,7 +79,7 @@ void ll_printf(const LinkedList *const ll, void (*f)(const data*)) {
 	if(ll_len(ll) > 0)
 		f(ll_get(ll, 0));
 	data *item;
-	unsigned int i;
+	size_t i;
 	for(i = 1; (item = ll_get(ll, i)); ++i) {
 			printf(", ");
 			f(item);
